Moves file to utils
